requirements.txt
mysql-connector-python
boto3
